feat: restructured code
