+++ Day7n8n9/github-api-project-ts/src/components/index.ts
@@ -1,5 +1,7 @@
 import Card from "./Card/Card";
 import ShowButton from "./Buttons/ShowButton/ShowButton";
 import Loader from "./Loader/Loader";
+import ErrorPage from "./ErrorPage/ErrorPage";
+import SelectByCity from "./SelectByCity/SelectByCity";
 
-export { Card, Loader, ShowButton };
+export { Card, Loader, ShowButton, ErrorPage, SelectByCity };
