+++ Day7n8n9/github-api-project-ts/src/containers/SearchGithubUsers/SearchGithubUsers.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+
+import { SelectByCity } from "../../components";
+
+type Props = {};
+
+const SearchGithubUsers = (props: Props) => {
+  return (
+    <div>
+      <SelectByCity />
+    </div>
+  );
+};
+
+export default SearchGithubUsers;
