+++ Day7n8n9/github-api-project-ts/src/containers/index.ts
@@ -1,5 +1,7 @@
 import Header from "./Header/Header";
 import Navbar from "./NavBar/Navbar";
 import Members from "./Members/Members";
+import SearchGithubUsers from "./SearchGithubUsers/SearchGithubUsers";
+import MemberInfo from "./MemberInfo/MemberInfo";
 
-export { Header, Navbar, Members };
+export { Header, Navbar, Members, SearchGithubUsers, MemberInfo };
