+++ Day7n8n9/github-api-project-ts/src/index.tsx
@@ -1,42 +1,19 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
+import { RouterProvider } from "react-router-dom";
+import reportWebVitals from "./reportWebVitals";
 
 import "./index.css";
-import reportWebVitals from "./reportWebVitals";
-import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import { HomePage } from "./pages";
-import { ShowButton } from "./components";
-import { Members } from "./containers";
-import MemberInfo from "./containers/MemberInfo/MemberInfo";
-import ErrorPage from "./components/ErrorPage/ErrorPage";
+import { myAppRouter } from "./routes";
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 
-const myAppRouter = createBrowserRouter([
-  {
-    path: "/",
-    element: <HomePage />,
-    errorElement: <ErrorPage />,
-    children: [
-      { index: true, element: <ShowButton /> },
-      {
-        path: "/members",
-        element: <Members />,
-      },
-      {
-        path: "/members/:id",
-        element: <MemberInfo />,
-      },
-    ],
-  },
-]);
-
 root.render(
-  // <React.StrictMode>
-  // </React.StrictMode>
-  <RouterProvider router={myAppRouter} />
+  <React.StrictMode>
+    <RouterProvider router={myAppRouter} />
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
