+++ Day7n8n9/github-api-project-ts/src/routes/index.tsx
@@ -1,0 +1,28 @@
+import { createBrowserRouter } from "react-router-dom";
+
+import { HomePage } from "../pages";
+import { ShowButton, ErrorPage } from "../components";
+import { Members, SearchGithubUsers, MemberInfo } from "../containers";
+
+export const myAppRouter = createBrowserRouter([
+  {
+    path: "/",
+    element: <HomePage />,
+    errorElement: <ErrorPage />,
+    children: [
+      { index: true, element: <ShowButton buttonText='Show' /> },
+      {
+        path: "/members",
+        element: <Members />,
+      },
+      {
+        path: "/search",
+        element: <SearchGithubUsers />,
+      },
+      {
+        path: "/members/:id",
+        element: <MemberInfo />,
+      },
+    ],
+  },
+]);
